Move trades-to-ohlcv to converter file
